-- Re-replicate to recover a corrupted chunk   -- API/tool to change the degree of replication on the fly   -- Change the client library code to push data for writes to multiple servers concurrently   -- Add support for O_EXCL in create (client/metaserver)   -- In the metaserver, maintain a list of chunks for which we need to check for       re-replication, as opposed to checking all the chunks (performance optimization)   -- Move the libkfs code to namespace

+++ src/cc/chunk/ChunkManager.cc
@@ -29,14 +29,17 @@ extern "C" {
 #include <sys/resource.h>
 #include <unistd.h>
 }
-#include "ChunkManager.h"
+
 #include "common/log.h"
-#include "Utils.h"
 #include "common/kfstypes.h"
+
+#include "ChunkManager.h"
+#include "MetaServerSM.h"
+#include "Utils.h"
+
 #include "libkfsIO/Counter.h"
 #include "libkfsIO/Checksum.h"
 #include "libkfsIO/Globals.h"
-using namespace libkfsio;
 
 #include <fstream>
 #include <sstream>
@@ -48,13 +51,18 @@ using namespace libkfsio;
 using std::ofstream;
 using std::ifstream;
 using std::istringstream;
+using std::ostringstream;
 using std::ios_base;
+using std::list;
 using std::min;
 using std::max;
 using std::endl;
 using std::find_if;
 using std::string;
 using std::vector;
+
+using namespace KFS;
+using namespace KFS::libkfsio;
 
 // Cleanup fds on which no I/O has been done for the past N secs
 const int INACTIVE_FDS_CLEANUP_INTERVAL_SECS = 300;
@@ -110,7 +118,7 @@ ChunkManager::AllocChunk(kfsFileId_t fileId, kfsChunkId_t chunkId,
 
     tableEntry = mChunkTable.find(chunkId);
     if (tableEntry != mChunkTable.end()) {
-        COSMIX_LOG_DEBUG("Chunk %s already exists; changing version # to %ld",
+        KFS_LOG_DEBUG("Chunk %s already exists; changing version # to %ld",
                          chunkPathname.c_str(), chunkVersion);
         cih = tableEntry->second;
         cih->chunkInfo.chunkVersion = chunkVersion;
@@ -118,7 +126,7 @@ ChunkManager::AllocChunk(kfsFileId_t fileId, kfsChunkId_t chunkId,
         return 0;
     }
     
-    COSMIX_LOG_DEBUG("Creating chunk: %s", chunkPathname.c_str());
+    KFS_LOG_DEBUG("Creating chunk: %s", chunkPathname.c_str());
 
     CleanupInactiveFds();
 
@@ -179,7 +187,7 @@ ChunkManager::StaleChunk(kfsChunkId_t chunkId)
     
     rename(chunkPathname.c_str(), staleChunkPathname.c_str());
 
-    COSMIX_LOG_INFO("Moving chunk %ld to staleChunks dir", chunkId);
+    KFS_LOG_INFO("Moving chunk %ld to staleChunks dir", chunkId);
 
     cih = tableEntry->second;
     mUsedSpace -= cih->chunkInfo.chunkSize;
@@ -245,7 +253,7 @@ ChunkManager::ChangeChunkVers(kfsFileId_t fileId,
 
     mIsChunkTableDirty = true;
 
-    COSMIX_LOG_DEBUG("Chunk %s already exists; changing version # to %ld",
+    KFS_LOG_DEBUG("Chunk %s already exists; changing version # to %ld",
                      chunkPathname.c_str(), chunkVersion);
 
     cih = tableEntry->second;
@@ -267,7 +275,7 @@ ChunkManager::ReplicationDone(kfsChunkId_t chunkId)
 
 #ifdef DEBUG
     string chunkPathname = MakeChunkPathname(chunkId);
-    COSMIX_LOG_DEBUG("Replication for chunk %s is complete...",
+    KFS_LOG_DEBUG("Replication for chunk %s is complete...",
                      chunkPathname.c_str());
 #endif
 
@@ -306,7 +314,7 @@ ChunkManager::MakeStaleChunkPathname(kfsChunkId_t chunkId)
 {
     ostringstream os;
 
-    os << mChunkBaseDir << "/stalechunks/" << chunkId;
+    os << mChunkBaseDir << "/lost+found/" << chunkId;
     return os.str();
 }
 
@@ -322,7 +330,7 @@ ChunkManager::OpenChunk(kfsChunkId_t chunkId,
     chunkPathname = MakeChunkPathname(chunkId);
     tableEntry = mChunkTable.find(chunkId);
     if (tableEntry == mChunkTable.end()) {
-        COSMIX_LOG_DEBUG("No such chunk: %s", chunkPathname.c_str());
+        KFS_LOG_DEBUG("No such chunk: %s", chunkPathname.c_str());
         return -EBADF;
     }
 
@@ -342,7 +350,7 @@ ChunkManager::OpenChunk(kfsChunkId_t chunkId,
         fd = cih->chunkHandle->mFileId;
     }
 
-    COSMIX_LOG_DEBUG("opening %s with flags %d; fd = %d", 
+    KFS_LOG_DEBUG("opening %s with flags %d; fd = %d", 
                      chunkPathname.c_str(), openFlags, fd);
 
     return 0;
@@ -366,7 +374,7 @@ ChunkManager::CloseChunk(kfsChunkId_t chunkId)
     if (cih->chunkHandle.use_count() <= 2) {
         if (cih->chunkHandle->mFileId < 0)
             return;
-        COSMIX_LOG_DEBUG("closing fileid = %d, for chunk = %ld",
+        KFS_LOG_DEBUG("closing fileid = %d, for chunk = %ld",
                          cih->chunkHandle->mFileId, 
                          cih->chunkHandle->mChunkId);
         close(cih->chunkHandle->mFileId);
@@ -406,7 +414,7 @@ ChunkManager::ReadChunk(ReadOp *op)
         return -KFS::ESERVERBUSY;
 
     if (op->chunkVersion != cih->chunkInfo.chunkVersion) {
-        COSMIX_LOG_DEBUG("Version # mismatch(have=%u vs asked=%ld...failing a read",
+        KFS_LOG_DEBUG("Version # mismatch(have=%u vs asked=%ld...failing a read",
                          cih->chunkInfo.chunkVersion, op->chunkVersion);
         return -KFS::EBADVERS;
     }
@@ -489,7 +497,7 @@ ChunkManager::WriteChunk(WriteOp *op)
             // issue a read
             ReadOp *rop = new ReadOp(op, OffsetToChecksumBlockStart(op->offset),
                                      CHECKSUM_BLOCKSIZE);
-            COSMIX_LOG_DEBUG("Write triggered a read for offset = %ld",
+            KFS_LOG_DEBUG("Write triggered a read for offset = %ld",
                              op->offset);
             rop->Execute();
 
@@ -546,7 +554,7 @@ ChunkManager::WriteChunk(WriteOp *op)
 
     op->diskConnection.reset(d);
 
-    COSMIX_LOG_DEBUG("Checksum for chunk: %ld, offset = %ld, bytes = %ld, cksum = %u",
+    KFS_LOG_DEBUG("Checksum for chunk: %ld, offset = %ld, bytes = %ld, cksum = %u",
                      op->chunkId, op->offset, op->numBytesIO, op->checksums[0]);
 
     return op->diskConnection->Write(op->offset, op->numBytesIO, op->dataBuf);
@@ -591,7 +599,7 @@ ChunkManager::ReadChunkDone(ReadOp *op)
         (op->chunkVersion != cih->chunkInfo.chunkVersion)) {
         AdjustDataRead(op);
         if (cih) {
-            COSMIX_LOG_DEBUG("Version # mismatch(have=%u vs asked=%ld...",
+            KFS_LOG_DEBUG("Version # mismatch(have=%u vs asked=%ld...",
                              cih->chunkInfo.chunkVersion, op->chunkVersion);
         }
         op->status = -KFS::EBADVERS;
@@ -615,13 +623,39 @@ ChunkManager::ReadChunkDone(ReadOp *op)
         return;
     }
 
-    COSMIX_LOG_DEBUG("Checksum mismatch for chunk=%ld, offset=%ld, bytes = %ld: expect: %u, computed: %u ",
-                     op->chunkId, op->offset, op->numBytesIO,
-                     cih->chunkInfo.chunkBlockChecksum[checksumBlock],
-                     checksum);
-    die("Checksum mismatch");
+    KFS_LOG_ERROR("Checksum mismatch for chunk=%ld, offset=%ld, bytes = %ld: expect: %u, computed: %u ",
+                  op->chunkId, op->offset, op->numBytesIO,
+                  cih->chunkInfo.chunkBlockChecksum[checksumBlock],
+                  checksum);
 
     op->status = -KFS::EBADCKSUM;
+
+    // Notify the metaserver that the chunk we have is "bad"; the
+    // metaserver will re-replicate this chunk.
+    NotifyMetaCorruptedChunk(op->chunkId);
+    
+    // Take out the chunk from our side
+    StaleChunk(op->chunkId);
+}
+
+void
+ChunkManager::NotifyMetaCorruptedChunk(kfsChunkId_t chunkId)
+{
+    ChunkInfoHandle_t *cih;
+
+    if (GetChunkInfoHandle(chunkId, &cih) < 0) {
+        KFS_LOG_ERROR("Unable to notify metaserver of corrupt chunk: %ld",
+                      chunkId);
+        return;
+    }
+
+    KFS_LOG_INFO("Notifying metaserver of corrupt chunk (%ld) in file %ld",
+                 cih->chunkInfo.fileId, chunkId);
+
+    // This op will get deleted when we get an ack from the metaserver
+    CorruptChunkOp *ccop = new CorruptChunkOp(0, cih->chunkInfo.fileId, 
+                                              chunkId);
+    gMetaServerSM.SubmitOp(ccop);
 }
 
 void
@@ -703,7 +737,7 @@ ChunkManager::Checkpoint()
     if (!mIsChunkTableDirty)
         return;
 
-    COSMIX_LOG_DEBUG("Checkpointing state");
+    KFS_LOG_DEBUG("Checkpointing state");
     cop = new CheckpointOp(1);
     
     for (iter = mChunkTable.begin(); iter != mChunkTable.end(); ++iter) {
@@ -755,12 +789,11 @@ ChunkManager::Restart()
     // sort all the chunk names alphabetically
     numChunkFiles = scandir(mChunkBaseDir, &namelist, 0, alphasort);
     if (numChunkFiles < 0) {
-        COSMIX_LOG_INFO("Unable to open %s", mChunkBaseDir);
+        KFS_LOG_INFO("Unable to open %s", mChunkBaseDir);
         return;
     }
 
     gLogger.Restore();
-
 
     // Now, validate: for each entry in the chunk table, verify that
     // the backing file exists. also, if there any "zombies" lying
@@ -783,7 +816,7 @@ ChunkManager::Restart()
             }
         }
         if (!found) {
-            COSMIX_LOG_DEBUG("Orphaned chunk as the file doesn't exist: %s",
+            KFS_LOG_DEBUG("Orphaned chunk as the file doesn't exist: %s",
                              chunkIdStr.c_str());
             orphans.push_back(entry.chunkId);
             continue;
@@ -796,7 +829,7 @@ ChunkManager::Restart()
 
         // stat buf's st_size is of type off_t.  Typecast to avoid compiler warnings.
         if (buf.st_size != (off_t) entry.chunkSize) {
-            COSMIX_LOG_DEBUG("Truncating file: %s to size: %zd",
+            KFS_LOG_DEBUG("Truncating file: %s to size: %zd",
                              chunkPathname.c_str(), entry.chunkSize);
             if (truncate(chunkPathname.c_str(), entry.chunkSize) < 0) {
                 perror("Truncate");
@@ -812,7 +845,7 @@ ChunkManager::Restart()
     // Get rid of the orphans---valid entries but no backing file
     for (j = 0; j < orphans.size(); ++j) {
 
-        COSMIX_LOG_DEBUG("Found orphan entry: %ld",
+        KFS_LOG_DEBUG("Found orphan entry: %ld",
                          orphans[j]);
 
         iter = mChunkTable.find(orphans[j]);
@@ -838,11 +871,16 @@ ChunkManager::Restart()
 
         // zombie
         chunkPathname = MakeChunkPathname(namelist[i]->d_name);
-        unlink(chunkPathname.c_str());
+
+        // there could be directories here...such as lost+found etc...
+        res = stat(chunkPathname.c_str(), &buf);
+        if ((res == 0) && (S_ISREG(buf.st_mode))) {
+            unlink(chunkPathname.c_str());
+            KFS_LOG_DEBUG("Found zombie entry: %s", chunkPathname.c_str());
+        }
+
         free(namelist[i]);
         namelist[i] = NULL;
-
-        COSMIX_LOG_DEBUG("Found zombie entry: %s", chunkPathname.c_str());
     }
 
     free(namelist);
@@ -896,7 +934,7 @@ ChunkManager::ReplayChangeChunkVers(kfsFileId_t fileId, kfsChunkId_t chunkId,
     if (GetChunkInfoHandle(chunkId, &cih) != 0) 
         return;
 
-    COSMIX_LOG_DEBUG("Chunk %ld already exists; changing version # to %ld",
+    KFS_LOG_DEBUG("Chunk %ld already exists; changing version # to %ld",
                      chunkId, chunkVersion);
     
     // Update the version #
@@ -1008,7 +1046,7 @@ ChunkManager::EnqueueWrite(WritePrepareOp *wp)
         return -EBADF;
 
     if (wp->chunkVersion != cih->chunkInfo.chunkVersion) {
-        COSMIX_LOG_DEBUG("Version # mismatch(have=%d vs asked=%lu...failing a write",
+        KFS_LOG_DEBUG("Version # mismatch(have=%d vs asked=%lu...failing a write",
                          cih->chunkInfo.chunkVersion, wp->chunkVersion);
         return -EINVAL;
     }
@@ -1097,7 +1135,7 @@ ChunkManager::ScavengePendingWrites()
             break;
         }
         // if it exceeds 5 mins, retire the op
-        COSMIX_LOG_DEBUG("Retiring write with id=%ld as it has been too long",
+        KFS_LOG_DEBUG("Retiring write with id=%ld as it has been too long",
                          op->writeId);
         mPendingWrites.pop_front();
         CloseChunk(op->chunkId);
@@ -1127,7 +1165,7 @@ public:
             return;
 
         // we have a valid file-id and it has been over 5 mins since we last did I/O on it.
-        COSMIX_LOG_DEBUG("cleanup: closing fileid = %d, for chunk = %ld",
+        KFS_LOG_DEBUG("cleanup: closing fileid = %d, for chunk = %ld",
                          cih->chunkHandle->mFileId, 
                          cih->chunkHandle->mChunkId);
         close(cih->chunkHandle->mFileId);
@@ -1154,7 +1192,7 @@ ChunkManager::CleanupInactiveFds()
             // bump the soft limit to the hard limit
             rlim.rlim_cur = rlim.rlim_max;
             if (setrlimit(RLIMIT_NOFILE, &rlim) == 0) {
-                COSMIX_LOG_DEBUG("Setting # of open files to: %ld",
+                KFS_LOG_DEBUG("Setting # of open files to: %ld",
                                  rlim.rlim_cur);
                 OPEN_FDS_LOW_WATERMARK = rlim.rlim_cur / 2;
             }
